pyproject.toml
[project]
name = "tap-googleads"
version = "0.13.0"
description = "`tap-googleads` is a Singer tap for GoogleAds, built with the Meltano SDK for Singer Taps."
authors = [
    { name = "AutoIDM" },
    { name = "Matatika" },
]
requires-python = ">=3.8,<4"
license = "Apache-2.0"
keywords = [
    "ELT",
    "GoogleAds",
]
dependencies = [
    "requests>=2.25.1,<3",
    "singer-sdk==0.42.1",
    "sqlparse>=0.5.3,<0.6",
    "pyhumps>=3.8.0,<4",
    "google-auth>=2.16.0,<3",
]

[project.scripts]
tap-googleads = "tap_googleads.tap:TapGoogleAds.cli"

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9",
    "tox>=4.25.0,<5",
    "flake8>=5.0.4,<6",
    "black~=24.8",
    "pydocstyle>=6.1.1,<7",
    "mypy~=1.14",
    "types-requests>=2.32.0,<3",
    "responses==0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
